# Speedup x 1.85: Generate `__version__` at build to avoid slow `importlib.metadata` import