--- a/.pre-commit-config.yaml
+++ b/.pre-commit-config.yaml
@@ -36,14 +36,6 @@ repos:
     hooks:
       - id: actionlint
 
-  - repo: https://github.com/pre-commit/mirrors-mypy
-    rev: v1.11.2
-    hooks:
-      - id: mypy
-        additional_dependencies: [pytest, types-freezegun, types-setuptools]
-        args: [--strict, --pretty, --show-error-codes, .]
-        pass_filenames: false
-
   - repo: https://github.com/tox-dev/pyproject-fmt
     rev: 2.2.3
     hooks: