--- a/tox.ini
+++ b/tox.ini
@@ -4,6 +4,7 @@ requires =
 env_list =
     docs
     lint
+    mypy
     py{py3, 313, 312, 311, 310, 39, 38}
 
 [testenv]
@@ -36,6 +37,15 @@ pass_env =
     PRE_COMMIT_COLOR
 commands =
     pre-commit run --all-files --show-diff-on-failure
+
+[testenv:mypy]
+deps =
+    mypy==1.11.2
+    pytest
+    types-freezegun
+    types-setuptools
+commands =
+    mypy --strict --pretty --show-error-codes . {posargs}
 
 [pytest]
 addopts = --color=yes --doctest-modules