Like https://github.com/hugovk/tinytext/pull/195.

With Python 3.13.0rc2:

```sh
python -X importtime -c "import humanize" 2> import.log && tuna import.log
```

# `main`

<img width="1582" alt="image" src="https://github.com/user-attachments/assets/3facca4a-09ce-464e-9655-fcf3a19f05e2">

# PR

<img width="1582" alt="image" src="https://github.com/user-attachments/assets/4a513d4b-61ee-4d4b-bf18-392877681866">

# hyperfine

```console
❯ hyperfine --warmup 32 \
--prepare "git checkout rm-importlib.metadata" 'python3 -c "import humanize # PR"' \
--prepare "git checkout main"                  'python3 -c "import humanize # main"'
Benchmark 1: python3 -c "import humanize # PR"
  Time (mean ± σ):      23.5 ms ±   3.5 ms    [User: 18.0 ms, System: 4.3 ms]
  Range (min … max):    21.8 ms …  43.4 ms    75 runs

  Warning: Statistical outliers were detected. Consider re-running this benchmark on a quiet system without any interferences from other programs.

Benchmark 2: python3 -c "import humanize # main"
  Time (mean ± σ):      43.4 ms ±   2.3 ms    [User: 34.5 ms, System: 7.6 ms]
  Range (min … max):    41.8 ms …  52.9 ms    52 runs

  Warning: Statistical outliers were detected. Consider re-running this benchmark on a quiet system without any interferences from other programs.

Summary
  python3 -c "import humanize # PR" ran
    1.85 ± 0.29 times faster than python3 -c "import humanize # main"
```

---

Also move mypy from pre-commit to tox, so it runs when the version file has been generated.